Improve size_combinations scenario
- Use a custom tracer to ignore message traces in the settling time
- reduce number of configurations from ~5000 to ~3000
- simulate for at least a settling time and then a effective sim time
- improve progress output
